Add preliminary support for structural inheritance
This is an incomplete, experimental, work in progress start of work to
add support for structural inheritance to Smithy. There's much more work
to be done in figuring out how traits are resolved for different
integrations (e.g., HTTP bindings, event streams, and other knowledge
sources), and this is lacking any kind of specification for now.